Place field update 06/06/2019
New description for the assignment of place fields from the features matrix (fet data), processed by cluster and electrode. 

This does not aim to recreate the Antonis Basic script, but rather starts from scratch, with the aims to create a place field figure by comparing coordinates in space and the presence of spiking activity.